sentry-analytics-worker/requirements.txt
fastapi
uvicorn[standard]
duckdb
python-dotenv
